# Improve reply parser performance