The first byte in reply specifies reply type, e.g. '+' is a string.
Before the commit, parser was implemented using backtracking, but
it is completely unnecessary. Now we read the first byte and then
call a particular parser by hands.

The result of the patch is noticeable improvement of parser speed
and 8% decrease of allocations in the benchmark. Before:

```
ping                  145971.76 Req/s
get                   133020.47 Req/s
mget                   94684.51 Req/s
ping (pipelined)     1007272.51 Req/s
multiExec get 1        93903.06 Req/s
multiExec get 50      560180.15 Req/s
multiExec get 1000    572504.88 Req/s
   6,115,428,248 bytes allocated in the heap
     209,206,824 bytes copied during GC
       6,261,232 bytes maximum residency (48 sample(s))
         821,264 bytes maximum slop
              15 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max
pause
  Gen  0      8561 colls,     0 par    0.100s   0.203s     0.0000s
0.0003s
  Gen  1        48 colls,     0 par    0.024s   0.030s     0.0006s
0.0034s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time    1.764s  (  3.788s elapsed)
  GC      time    0.124s  (  0.232s elapsed)
  EXIT    time    0.000s  (  0.000s elapsed)
  Total   time    1.888s  (  4.021s elapsed)

  %GC     time       6.6%  (5.8% elapsed)

  Alloc rate    3,466,796,058 bytes per MUT second

  Productivity  93.4% of total user, 43.9% of total elapsed
```

After:

```
ping                  154106.71 Req/s
get                   133463.24 Req/s
mget                   97220.66 Req/s
ping (pipelined)     1066245.85 Req/s
multiExec get 1        97206.76 Req/s
multiExec get 50      601576.13 Req/s
multiExec get 1000    487369.81 Req/s
   5,627,326,864 bytes allocated in the heap
     209,870,776 bytes copied during GC
       6,268,320 bytes maximum residency (45 sample(s))
         746,968 bytes maximum slop
              16 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max
pause
  Gen  0      7677 colls,     0 par    0.112s   0.208s     0.0000s
0.0005s
  Gen  1        45 colls,     0 par    0.040s   0.037s     0.0008s
0.0072s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time    1.776s  (  3.684s elapsed)
  GC      time    0.152s  (  0.245s elapsed)
  EXIT    time    0.000s  (  0.000s elapsed)
  Total   time    1.928s  (  3.930s elapsed)

  %GC     time       7.9%  (6.2% elapsed)

  Alloc rate    3,168,539,900 bytes per MUT second

  Productivity  92.1% of total user, 45.2% of total elapsed
```