@qrilka 2016-03-26
And I wonder why `multiExec get 1000` showed 15% lesser value :-\