@Yuras 2016-03-26
@qrilka It is just a random variation (GC? network glitch?). Here is an other run with the patch applied:

```
ping                  150917.20 Req/s
get                   131392.75 Req/s
mget                   96057.14 Req/s
ping (pipelined)     1045970.40 Req/s
multiExec get 1        95334.25 Req/s
multiExec get 50      591996.21 Req/s
multiExec get 1000    603864.73 Req/s
   5,610,688,232 bytes allocated in the heap
     213,042,032 bytes copied during GC
       5,905,912 bytes maximum residency (47 sample(s))
         832,744 bytes maximum slop
              15 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0      7665 colls,     0 par    0.108s   0.201s     0.0000s    0.0003s
  Gen  1        47 colls,     0 par    0.016s   0.028s     0.0006s    0.0023s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time    1.716s  (  3.719s elapsed)
  GC      time    0.124s  (  0.230s elapsed)
  EXIT    time    0.000s  (  0.000s elapsed)
  Total   time    1.840s  (  3.949s elapsed)

  %GC     time       6.7%  (5.8% elapsed)

  Alloc rate    3,269,631,836 bytes per MUT second

  Productivity  93.3% of total user, 43.5% of total elapsed
```

Basically all benchmarks were improved on average. Decrease of allocations is more reproducible though.