@informatikr 2016-03-26
Nice performance improvement, thanks!

The `multiExec get 1000` test often produces (slower) outliers. I think this is because

1. during the multiExec it builds up 1000 `Queued` futures and
2. the final EXEC returns a large reply that needs to be parsed and put into a 1000 element vector.

These two mean more memory use and more GC compared to the smaller tests. So I agree that this is a random variation.